[C-1615] Fix android fullscreen modals
