+++ packages/mobile/android/app/src/main/res/values-night/colors.xml
@@ -1,0 +1,3 @@
+<resources>
+    <color name="colorPrimary">#C74BD3</color>
+</resources>
