+++ packages/mobile/src/screens/app-screen/AppScreen.tsx
@@ -1,55 +1,39 @@
-import { useEffect } from 'react'
+import { createNativeStackNavigator } from '@react-navigation/native-stack'
 
-import { walletActions } from '@audius/common'
-import { useAppState } from '@react-native-community/hooks'
-import type { BottomTabBarProps } from '@react-navigation/bottom-tabs'
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
-import type { NavigatorScreenParams } from '@react-navigation/native'
-import { useDispatch } from 'react-redux'
+import { EditExistingTrackScreen } from '../edit-track-screen'
+import { TipArtistModal } from '../tip-artist-screen'
+import { UploadScreen } from '../upload-screen'
+import { WalletConnectScreen } from '../wallet-connect'
 
-import { AppTabBar } from './AppTabBar'
-import type { ExploreTabScreenParamList } from './ExploreTabScreen'
-import { ExploreTabScreen } from './ExploreTabScreen'
-import type { FavoritesTabScreenParamList } from './FavoritesTabScreen'
-import { FavoritesTabScreen } from './FavoritesTabScreen'
-import type { FeedTabScreenParamList } from './FeedTabScreen'
-import { FeedTabScreen } from './FeedTabScreen'
-import { NotificationsTabScreen } from './NotificationsTabScreen'
-import type { ProfileTabScreenParamList } from './ProfileTabScreen'
-import type { TrendingTabScreenParamList } from './TrendingTabScreen'
-import { TrendingTabScreen } from './TrendingTabScreen'
-const { getBalance } = walletActions
+import { AppTabsScreen } from './AppTabsScreen'
 
-export type AppScreenParamList = {
-  feed: NavigatorScreenParams<FeedTabScreenParamList>
-  trending: NavigatorScreenParams<TrendingTabScreenParamList>
-  explore: NavigatorScreenParams<ExploreTabScreenParamList>
-  favorites: NavigatorScreenParams<FavoritesTabScreenParamList>
-  profile: NavigatorScreenParams<ProfileTabScreenParamList>
-}
-
-const Tab = createBottomTabNavigator()
-
-const screenOptions = { headerShown: false }
-const tabBar = (props: BottomTabBarProps) => <AppTabBar {...props} />
+const Stack = createNativeStackNavigator()
+const modalScreenOptions = { presentation: 'fullScreenModal' as const }
 
 export const AppScreen = () => {
-  const dispatch = useDispatch()
-  const appState = useAppState()
-
-  useEffect(() => {
-    if (appState === 'active') {
-      dispatch(getBalance())
-    }
-  }, [appState, dispatch])
-
   return (
-    <Tab.Navigator tabBar={tabBar} screenOptions={screenOptions}>
-      <Tab.Screen name='feed' component={FeedTabScreen} />
-      <Tab.Screen name='trending' component={TrendingTabScreen} />
-      <Tab.Screen name='explore' component={ExploreTabScreen} />
-      <Tab.Screen name='favorites' component={FavoritesTabScreen} />
-      <Tab.Screen name='notifications' component={NotificationsTabScreen} />
-    </Tab.Navigator>
+    <Stack.Navigator screenOptions={{ headerShown: false }}>
+      <Stack.Screen name='AppTabs' component={AppTabsScreen} />
+      <Stack.Screen
+        name='TipArtist'
+        component={TipArtistModal}
+        options={modalScreenOptions}
+      />
+      <Stack.Screen
+        name='Upload'
+        component={UploadScreen}
+        options={modalScreenOptions}
+      />
+      <Stack.Screen
+        name='EditTrack'
+        component={EditExistingTrackScreen}
+        options={modalScreenOptions}
+      />
+      <Stack.Screen
+        name='WalletConnect'
+        component={WalletConnectScreen}
+        options={modalScreenOptions}
+      />
+    </Stack.Navigator>
   )
 }
