+++ packages/mobile/src/screens/app-screen/index.ts
@@ -1,4 +1,5 @@
 export * from './AppScreen'
+export * from './AppTabsScreen'
 export * from './AppTabScreen'
 export * from './TopBarIconButton'
 export * from './AppTabNavigationProvider'
